Make Handle close on drop and Send + Sync
You now must construct a handle by calling the unsafe function
`Handle::from_raw`, which requires that you guarantee that the
underlying HANDLE is thread safe.

It's worth noting that most handles in winapi are thread safe.

`Handle`s also needs to be shared currently, so we wrap the inner
handle in an Arc.

We also remove all conversion APIs which allows constructing a `Handle`
from a raw `HANDLE`, since this could be used to circumvent the
guarantees provided by `Send` and `Sync` without using `unsafe`.

This also adds bookkeeping for closing handles. Since we didn't do this
before, we were leaking them instead on every operation.